backend/requirements.txt
flask
tensorflow
numpy
sklearn.preprocessing
joblib
json